Tidy up content handlers
